Move compare_* test methods to test module
