+++ cfplot/cfplot.py
@@ -11374,135 +11374,3 @@ def _process_color_scales():
         print(div)
         print("")
         print("")
-
-
-# ====================================================
-# TESTING - to be moved out to dedicated test module
-# ====================================================
-
-def compare_images(example=None):
-    """
-    Compare images and return an error string if they don't match.
-    """
-    import hashlib
-
-    # TODO SLB: convert all 'home/andy/' paths to general configurable path
-    # (there are many examples below but also throughout this script)
-
-    disp = _which("display")
-    conv = _which("convert")
-    comp = _which("compare")
-    file = f"fig{example}.png"
-    file_new = f"/home/andy/cfplot.src/cfplot/{file}"
-    file_ref = f"/home/andy/regression/{file}"
-
-    # Check md5 checksums are the same and display files if not
-    if (
-        hashlib.md5(open(file_new, "rb").read()).hexdigest()
-        != hashlib.md5(open(file_ref, "rb").read()).hexdigest()
-    ):
-        print(f"***Failed example {example}***")
-        error_image = f"/home/andy/cfplot.src/cfplot/error_{file}"
-        diff_image = f"/home/andy/cfplot.src/cfplot/difference_{file}"
-        p = subprocess.Popen([comp, file_new, file_ref, diff_image])
-        (output, err) = p.communicate()
-        p.wait()
-        p = subprocess.Popen(
-            [conv, "+append", file_new, file_ref, error_image]
-        )
-        (output, err) = p.communicate()
-        p.wait()
-        subprocess.Popen([disp, diff_image])
-
-    else:
-        print(f"Passed example {example}")
-
-
-def compare_arrays(
-    ref=None,
-    levs_test=None,
-    gvals_test=None,
-    mapaxis_test=None,
-    min=None,
-    max=None,
-    step=None,
-    mult=None,
-    type=None,
-):
-    """
-    Compare arrays and return an error string if they don't match.
-    """
-
-    anom = 0
-    if levs_test:
-        levs(min, max, step)
-        if np.size(ref) != np.size(plotvars.levels):
-            anom = 1
-        else:
-            for val in np.arange(np.size(ref)):
-                if abs(ref[val] - plotvars.levels[val]) >= 1e-6:
-                    anom = 1
-
-        if anom == 1:
-            print(
-                "***levs failure***\n"
-                f"min, max, step are {min}, {max}, {step}\n"
-                "generated levels are:\n"
-                f"{plotvars.levels}\n"
-                f"expected levels:\n{ref}"
-            )
-        else:
-            pass_str = f"Passed cfp.levs(min={min}, max={max}, step={step})"
-            print(pass_str)
-
-    anom = 0
-    if gvals_test:
-        vals, testmult = _gvals(min, max)
-        if np.size(ref) != np.size(vals):
-            anom = 1
-        else:
-            for val in np.arange(np.size(ref)):
-                if abs(ref[val] - vals[val]) >= 1e-6:
-                    anom = 1
-        if mult != testmult:
-            anom = 1
-
-        if anom == 1:
-            print(
-                "***gvals failure***\n"
-                f"cfp._gvals({min}, {max})\n\n"
-                f"generated values are:{vals}\n"
-                f"with a  multiplier of {testmult}\n\n"
-                f"expected values:{ref}\n"
-                f"with a multiplier of {mult}\n"
-            )
-        else:
-            pass_str = f"Passed cfp._gvals({min}, {max})"
-            print(pass_str)
-
-    anom = 0
-    if mapaxis_test:
-        ref_ticks = ref[0]
-        ref_labels = ref[1]
-        test_ticks, test_labels = _mapaxis(min=min, max=max, type=type)
-        if np.size(test_ticks) != np.size(ref_ticks):
-            anom = 1
-        else:
-            for val in np.arange(np.size(ref_ticks)):
-                if abs(ref_ticks[val] - test_ticks[val]) >= 1e-6:
-                    anom = 1
-                if ref_labels[val] != test_labels[val]:
-                    anom = 1
-
-        if anom == 1:
-            print(
-                "***mapaxis failure***\n\n"
-                f"cfp._mapaxis(min={min}, max={max}, type={type})\n"
-                f"generated values are:{test_ticks}\n"
-                f"with labels:{test_labels}\n\n"
-                f"expected ticks:{ref_ticks}\n"
-                f"with labels:{ref_labels}\n"
-            )
-        else:
-            pass_str = f"Passed cfp._mapaxis<(min={min}, max={max}, type={type})"
-            print(pass_str)
